Mode DAO changes
mode has now an Int id
single<int> is returned to avoid blocking

+++ Android/app/src/main/kotlin/com/onewisebit/scpescape/model/entities/Game.kt
@@ -17,7 +17,7 @@ data class Game(
     @ColumnInfo(name = "game_ID")
     val id: Long,
     @ColumnInfo(name = "mode")
-    val modeID: Long,
+    val modeID: Int,
     @ColumnInfo(name = "game_type")
     val type: Int,
     //TODO: test this value
